+++ lib/src/linear_optimizer/initial_coefficients_generator/zero_coefficients_generator.dart
@@ -1,5 +1,6 @@
 import 'package:ml_algo/src/linear_optimizer/initial_coefficients_generator/initial_coefficients_generator.dart';
-import 'package:ml_linalg/linalg.dart';
+import 'package:ml_linalg/dtype.dart';
+import 'package:ml_linalg/vector.dart';
 
 class ZeroCoefficientsGenerator implements InitialCoefficientsGenerator {
   ZeroCoefficientsGenerator(this.dtype);