+++ lib/src/linear_optimizer/least_squares_coordinate_descent_optimizer.dart
@@ -1,7 +1,9 @@
 import 'package:ml_algo/src/linear_optimizer/convergence_detector/convergence_detector.dart';
 import 'package:ml_algo/src/linear_optimizer/initial_coefficients_generator/initial_coefficients_generator.dart';
 import 'package:ml_algo/src/linear_optimizer/linear_optimizer.dart';
-import 'package:ml_linalg/linalg.dart';
+import 'package:ml_linalg/dtype.dart';
+import 'package:ml_linalg/matrix.dart';
+import 'package:ml_linalg/vector.dart';
 
 class LeastSquaresCoordinateDescentOptimizer implements LinearOptimizer {
   LeastSquaresCoordinateDescentOptimizer(