+++ lib/src/metric/regression/rmse.dart
@@ -2,7 +2,7 @@ import 'dart:math' as math;
 
 import 'package:ml_algo/src/helpers/validate_matrix_columns.dart';
 import 'package:ml_algo/src/metric/metric.dart';
-import 'package:ml_linalg/linalg.dart';
+import 'package:ml_linalg/matrix.dart';
 
 class RmseMetric implements Metric {
   const RmseMetric();