+++ test/regressor/knn_regressor/knn_regressor_integration_test.dart
@@ -1,14 +1,20 @@
 import 'dart:io';
 
-import 'package:ml_algo/ml_algo.dart';
 import 'package:ml_algo/src/common/constants/common_json_keys.dart';
+import 'package:ml_algo/src/knn_kernel/kernel_type.dart';
 import 'package:ml_algo/src/knn_kernel/kernel_type_json_keys.dart';
 import 'package:ml_algo/src/knn_solver/knn_solver_constants.dart';
 import 'package:ml_algo/src/knn_solver/knn_solver_json_keys.dart';
+import 'package:ml_algo/src/regressor/knn_regressor/knn_regressor.dart';
 import 'package:ml_algo/src/regressor/knn_regressor/knn_regressor_constants.dart';
 import 'package:ml_algo/src/regressor/knn_regressor/knn_regressor_json_keys.dart';
 import 'package:ml_dataframe/ml_dataframe.dart';
-import 'package:ml_linalg/linalg.dart';
+import 'package:ml_linalg/distance.dart';
+import 'package:ml_linalg/distance_type_to_json.dart';
+import 'package:ml_linalg/dtype.dart';
+import 'package:ml_linalg/dtype_to_json.dart';
+import 'package:ml_linalg/matrix.dart';
+import 'package:ml_linalg/matrix_to_json.dart';
 import 'package:test/test.dart';
 
 import '../../helpers.dart';