+++ test/retrieval/random_binary_projection_searcher/random_binary_projection_searcher_impl_test.dart
@@ -1,6 +1,6 @@
-import 'package:ml_algo/ml_algo.dart';
 import 'package:ml_algo/src/retrieval/random_binary_projection_searcher/random_binary_projection_searcher_impl.dart';
 import 'package:ml_dataframe/ml_dataframe.dart';
+import 'package:ml_linalg/distance.dart';
 import 'package:ml_linalg/vector.dart';
 import 'package:test/test.dart';